# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for ordering food.",
        "title": "Food Ordering API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/printers/categories/{id}": {
            "get": {
                "description": "ดึงรายการหมวดหมู่ที่เครื่องพิมพ์สามารถพิมพ์ได้",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Printer"
                ],
                "summary": "ดึงรายการหมวดหมู่ของเครื่องพิมพ์",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Printer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Category"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "กำหนดหมวดหมู่ที่เครื่องพิมพ์สามารถพิมพ์ได้",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Printer"
                ],
                "summary": "กำหนดหมวดหมู่ให้เครื่องพิมพ์",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Printer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "รายการ ID ของหมวดหมู่",
                        "name": "categories",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_handlers.AssignCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Printer"
                        }
                    }
                }
            }
        },
        "/api/printers/pending-jobs": {
            "get": {
                "description": "ดึงรายการงานพิมพ์ที่ยังไม่ได้พิมพ์สำหรับเครื่องพิมพ์ที่ระบุ (รองรับทั้ง IP และ USB)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Printer"
                ],
                "summary": "รับงานพิมพ์ที่รอดำเนินการ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Printer IP Address (สำหรับเครื่องพิมพ์เครือข่าย)",
                        "name": "printer_ip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Vendor ID (สำหรับเครื่องพิมพ์ USB)",
                        "name": "vendor_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Product ID (สำหรับเครื่องพิมพ์ USB)",
                        "name": "product_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PrintJob"
                            }
                        }
                    }
                }
            }
        },
        "/api/printers/status/{id}": {
            "put": {
                "description": "อัพเดทสถานะของงานพิมพ์ (completed/failed)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Printer"
                ],
                "summary": "อัพเดทสถานะงานพิมพ์",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Print Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New Status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PrintJob"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api_handlers.AssignCategoryRequest": {
            "type": "object",
            "required": [
                "category_ids"
            ],
            "properties": {
                "category_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "models.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nameCh": {
                    "type": "string"
                },
                "nameEn": {
                    "type": "string"
                }
            }
        },
...
        "models.Order": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderItem"
                    }
                },
                "receipt": {
                    "$ref": "#/definitions/models.Receipt"
                },
                "receiptID": {
                    "type": "integer"
                },
                "status": {
                    "description": "\"completed\", \"uncompleted\"",
                    "type": "string"
                },
                "tableID": {
                    "type": "integer"
                },
                "total": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "models.OrderItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menuItem": {
                    "$ref": "#/definitions/models.MenuItem"
                },
                "menuItemID": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "options": {
                    "description": "เพิ่ม relation กับ options",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderItemOption"
                    }
                },
                "order": {
                    "$ref": "#/definitions/models.Order"
                },
                "orderID": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "servedAt": {
                    "description": "เวลาที่เสิร์ฟอาหาร",
                    "type": "string"
                },
                "status": {
                    "description": "pending, served, cancelled",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.OrderItemOption": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menuOption": {
                    "$ref": "#/definitions/models.MenuOption"
                },
                "menuOptionID": {
                    "description": "Foreign key to MenuOption",
                    "type": "integer"
                },
                "orderItem": {
                    "$ref": "#/definitions/models.OrderItem"
                },
                "orderItemID": {
                    "description": "Foreign key to OrderItem",
                    "type": "integer"
                },
                "price": {
                    "description": "ราคา ณ เวลาที่สั่ง",
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "description": "ค่าที่เลือก เช่น \"เผ็ดมาก\", \"เพิ่มไข่ดาว\"",
                    "type": "string"
                }
            }
        },
        "models.PrintJob": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order": {
                    "$ref": "#/definitions/models.Order"
                },
                "orderID": {
                    "description": "nullable",
                    "type": "integer"
                },
                "printer": {
                    "$ref": "#/definitions/models.Printer"
                },
                "printerID": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Printer": {
            "type": "object",
            "properties": {
                "categories": {
                    "description": "หมวดหมู่ที่พิมพ์ได้",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Category"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "department": {
                    "description": "แผนก/ฝ่ายที่ใช้งาน",
                    "type": "string"
                },
                "description": {
                    "description": "รายละเอียดเพิ่มเติม",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ipaddress": {
                    "description": "IP Address",
                    "type": "string"
                },
                "lastSeen": {
                    "description": "เวลาที่เห็นเครื่องพิมพ์ล่าสุด",
                    "type": "string"
                },
                "name": {
                    "description": "ชื่อเครื่องพิมพ์",
                    "type": "string"
                },
                "port": {
                    "description": "Port number",
                    "type": "integer"
                },
                "productID": {
                    "description": "สำหรับ USB printer",
                    "type": "string"
                },
                "status": {
                    "description": "สถานะ: active, inactive, maintenance",
                    "type": "string"
                },
                "type": {
                    "description": "'network' หรือ 'usb'",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vendorID": {
                    "description": "สำหรับ USB printer",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}